Order capture patch summaries
